Move some helpers to SetupHelper class
